Создание модуляции для js
